Add files via upload
The weather app is an application that allows users to know the weather conditions in any country in the world. The application relies on an API (Application Programming Interface) to obtain weather data from external sources.

When the user selects a specific country in the application, the application sends a request to the associated API to obtain weather data in that country. The HTTP protocol is used to send the request and receive the response from the API.

The available weather data from the API includes temperature, wind speed, humidity, precipitation, and clouds. The application converts this data into easy-to-understand images and graphics that users can easily understand.

The application can be useful for users who want to travel to a specific country and want to know the weather conditions there, or for those who want to know the weather conditions in a specific country for business or personal purposes.